# backend/app/api/openapi.json
{
  "components": {
    "schemas": {
      "BaseRateFinding": {
        "description": "Historical hit rate for the closest matching setup bucket.",
        "properties": {
          "bucket": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Bucket",
            "type": "object"
          },
          "hit_rate": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Hit Rate",
            "type": "number"
          },
          "sample_size": {
            "minimum": 0.0,
            "title": "Sample Size",
            "type": "integer"
          },
          "setup": {
            "enum": [
              "orb",
              "sr_bounce",
              "other"
            ],
            "title": "Setup",
            "type": "string"
          }
        },
        "required": [
          "setup",
          "sample_size",
          "hit_rate"
        ],
        "title": "BaseRateFinding",
        "type": "object"
      },
      "BiasFinding": {
        "description": "One bias the bias detector identified in the thesis.",
        "properties": {
          "bias": {
            "enum": [
              "anchoring",
              "recency",
              "confirmation",
              "overconfidence",
              "revenge"
            ],
            "title": "Bias",
            "type": "string"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "severity": {
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Severity",
            "type": "string"
          },
          "span": {
            "description": "The phrase from the thesis that triggered the flag.",
            "title": "Span",
            "type": "string"
          }
        },
        "required": [
          "bias",
          "span",
          "severity",
          "explanation"
        ],
        "title": "BiasFinding",
        "type": "object"
      },
      "CritiquePending": {
        "description": "200 response from `GET /api/v1/critique/{id}` when work is still in flight.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "const": "pending",
            "default": "pending",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "created_at"
        ],
        "title": "CritiquePending",
        "type": "object"
      },
      "CritiqueQueued": {
        "description": "202 response from `POST /api/v1/critique`.",
        "properties": {
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "const": "queued",
            "default": "queued",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "request_id"
        ],
        "title": "CritiqueQueued",
        "type": "object"
      },
      "CritiqueRequest": {
        "description": "Submission body for `POST /api/v1/critique`.\n\n`thesis` is free-text English; the parser is responsible for extracting\nstructured fields. The ticker inside the thesis must carry an exchange\nindicator (`AAPL` / `INFY.NS` / `SAP.DE`); parser will fail closed otherwise.",
        "properties": {
          "thesis": {
            "maxLength": 4000,
            "minLength": 10,
            "title": "Thesis",
            "type": "string"
          }
        },
        "required": [
          "thesis"
        ],
        "title": "CritiqueRequest",
        "type": "object"
      },
      "CritiqueResponse": {
        "description": "Final critique body. Mirrors what the synthesizer produced and the critic verified.",
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "cost_usd": {
            "default": 0.0,
            "minimum": 0.0,
            "title": "Cost Usd",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "gap_flags": {
            "description": "Reasons a partial critique is partial; e.g. 'newsapi_unavailable'.",
            "items": {
              "type": "string"
            },
            "title": "Gap Flags",
            "type": "array"
          },
          "parsed_thesis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParsedThesis"
              },
              {
                "type": "null"
              }
            ]
          },
          "raw_thesis": {
            "title": "Raw Thesis",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "sections": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CritiqueSections"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "complete",
              "partial",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "verdict": {
            "anyOf": [
              {
                "enum": [
                  "strong",
                  "marginal",
                  "weak"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Top-line characterization. Absent on `failed`.",
            "title": "Verdict"
          }
        },
        "required": [
          "request_id",
          "status",
          "raw_thesis",
          "created_at"
        ],
        "title": "CritiqueResponse",
        "type": "object"
      },
      "CritiqueSections": {
        "description": "The four-section critique body, plus base-rate context.",
        "properties": {
          "base_rates": {
            "items": {
              "$ref": "#/components/schemas/BaseRateFinding"
            },
            "title": "Base Rates",
            "type": "array"
          },
          "bias": {
            "items": {
              "$ref": "#/components/schemas/BiasFinding"
            },
            "title": "Bias",
            "type": "array"
          },
          "disconfirming": {
            "items": {
              "$ref": "#/components/schemas/DisconfirmingItem"
            },
            "title": "Disconfirming",
            "type": "array"
          },
          "mechanics": {
            "$ref": "#/components/schemas/StructureFinding"
          },
          "setup_critique": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SetupCritique"
              },
              {
                "type": "null"
              }
            ]
          },
          "stress_test": {
            "items": {
              "$ref": "#/components/schemas/StressTestClaim"
            },
            "title": "Stress Test",
            "type": "array"
          }
        },
        "required": [
          "mechanics"
        ],
        "title": "CritiqueSections",
        "type": "object"
      },
      "DisconfirmingItem": {
        "description": "A piece of evidence that runs counter to the thesis.",
        "properties": {
          "headline": {
            "title": "Headline",
            "type": "string"
          },
          "relevance": {
            "enum": [
              "ticker",
              "sector",
              "catalyst",
              "peer"
            ],
            "title": "Relevance",
            "type": "string"
          },
          "source": {
            "description": "e.g. 'NewsAPI', 'YahooFinance RSS', 'EDGAR'.",
            "title": "Source",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "source",
          "headline",
          "relevance",
          "summary"
        ],
        "title": "DisconfirmingItem",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "description": "`GET /api/v1/health` body.",
        "properties": {
          "checks": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Per-dependency state: db, redis, last_llm.",
            "title": "Checks",
            "type": "object"
          },
          "status": {
            "enum": [
              "ok",
              "degraded",
              "down"
            ],
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "status",
          "version"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "ParsedThesis": {
        "description": "Structured form of the user's free-text thesis.",
        "properties": {
          "claims": {
            "description": "Verifiable claims for the stress-tester.",
            "items": {
              "type": "string"
            },
            "title": "Claims",
            "type": "array"
          },
          "direction": {
            "enum": [
              "long",
              "short"
            ],
            "title": "Direction",
            "type": "string"
          },
          "entry": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry"
          },
          "exchange": {
            "enum": [
              "US",
              "IN",
              "DE"
            ],
            "title": "Exchange",
            "type": "string"
          },
          "raw_confidence_words": {
            "description": "Phrases the bias detector pre-scan flagged.",
            "items": {
              "type": "string"
            },
            "title": "Raw Confidence Words",
            "type": "array"
          },
          "setup": {
            "enum": [
              "orb",
              "sr_bounce",
              "other"
            ],
            "title": "Setup",
            "type": "string"
          },
          "stop": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop"
          },
          "target": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          },
          "time_horizon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Horizon"
          }
        },
        "required": [
          "ticker",
          "exchange",
          "direction",
          "setup"
        ],
        "title": "ParsedThesis",
        "type": "object"
      },
      "SetupCritique": {
        "description": "Output of the setup specialist (ORB / S/R / generic).",
        "properties": {
          "checks": {
            "description": "Per-rule check rows, each {rule, status, evidence}.",
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Checks",
            "type": "array"
          },
          "narrative": {
            "title": "Narrative",
            "type": "string"
          },
          "quality": {
            "enum": [
              "strong",
              "marginal",
              "weak"
            ],
            "title": "Quality",
            "type": "string"
          },
          "setup": {
            "enum": [
              "orb",
              "sr_bounce",
              "other"
            ],
            "title": "Setup",
            "type": "string"
          }
        },
        "required": [
          "setup",
          "quality",
          "narrative"
        ],
        "title": "SetupCritique",
        "type": "object"
      },
      "StressTestClaim": {
        "description": "One claim from the thesis, with the stress-tester's verdict.",
        "properties": {
          "claim": {
            "title": "Claim",
            "type": "string"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "title": "Evidence",
            "type": "array"
          },
          "verdict": {
            "enum": [
              "supported",
              "contradicted",
              "unverifiable"
            ],
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "claim",
          "verdict"
        ],
        "title": "StressTestClaim",
        "type": "object"
      },
      "StructureFinding": {
        "description": "Output of the structure critic (trade mechanics).",
        "properties": {
          "atr_14": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atr 14"
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "title": "Issues",
            "type": "array"
          },
          "rr_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rr Ratio"
          },
          "sizing_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sizing Note"
          },
          "stop_to_atr_multiple": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop To Atr Multiple"
          },
          "verdict": {
            "description": "Prose summary of the mechanics critique.",
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "verdict"
        ],
        "title": "StructureFinding",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Trade-Idea Critic",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/critique": {
      "post": {
        "description": "Queue a critique. Returns the `request_id` the client polls or streams.",
        "operationId": "submit_critique_api_v1_critique_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CritiqueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CritiqueQueued"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit a thesis for critique",
        "tags": [
          "critique"
        ]
      }
    },
    "/api/v1/critique/{request_id}": {
      "get": {
        "description": "Return the critique body, or a `pending` envelope while work is in flight.",
        "operationId": "get_critique_api_v1_critique__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CritiqueResponse"
                    },
                    {
                      "$ref": "#/components/schemas/CritiquePending"
                    }
                  ],
                  "title": "Response Get Critique Api V1 Critique  Request Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Unknown request_id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fetch a critique",
        "tags": [
          "critique"
        ]
      }
    },
    "/api/v1/critique/{request_id}/stream": {
      "get": {
        "description": "Emit `status`, `node_started`, `node_completed`, `final` (or `error`) events.\n\nIf the run is already finished, a single `final` (or `error`) event is sent\nand the stream closes.",
        "operationId": "stream_critique_api_v1_critique__request_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "SSE stream. See docs/api/sse-events.md for event types."
          },
          "404": {
            "description": "Unknown request_id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream critique progress (SSE)",
        "tags": [
          "critique"
        ]
      }
    },
    "/api/v1/health": {
      "get": {
        "description": "Liveness probe. Phase 1.3 extends `checks` with real DB, Redis, last_llm states.",
        "operationId": "health_api_v1_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "meta"
        ]
      }
    }
  }
}